list: Provide a macro to init a static structure member that is a list
Provide a means to initialize a structure element that is a list that
happens to be global or static.  This can save us from having to do the
init at runtime (and thus not optimize away some code if the struct
isn't used at all).

So we now have METAL_DECLARE_LIST to define & init a global list named
foobar:

METAL_DECLARE_LIST(foobar);

and METAL_INIT_LIST to do:

struct foo {
        int a;
        struct metal_list x;
};

struct foo bar = {
        .a = 4,
        .x = METAL_INIT_LIST(bar.x),
};

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>