lib: xlnx: Use CMAKE_C_FLAGS to determine if SDT symbol is present
lib: generic: xlnx: Use CMAKE_C_FLAGS to determine if SDT symbol is
present.

Currently the file bspconfig.h file is used to determine if System Device
Tree workflow is being used. Instead change condition to be if
SDT is in CMAKE_C_FLAGS variable.

Acked-by: Sergei Korneichuk <sergei.korneichuk@amd.com>

lib: freertos: xlnx: Use CMAKE_C_FLAGS to determine if SDT symbol is
present.

Currently the file bspconfig.h file is used to determine if System Device
Tree workflow is being used. Instead change condition to be if SDT is in
CMAKE_C_FLAGS variable.

As part of this also make the freertos build up to parity with baremetal.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Reviewed-by: Tanmay Shah <tanmay.shah@amd.com>