examples: shmem_atomic_demo: prevent RPU from going ahead too early
There can be race condition where RPU side believes demo is done before
APU is actually finished. Then the RPU goes to the next demo. APU does
an IPI kick and RPU reads a buffer at the wrong time and believes it
is seeing corruption issue.

Fix this by adding a handshake so that:
- At the start of the RPU demo, the RPU sets a bit to denote it is waiting.
- At the start of the APU demo, set the same bit in memory so each party
  is in sync.
- Once RPU is done (which should be before APU is done), RPU will wait on
  the bit to flip.
- Once APU is done processing, it will flip the bit and both RPU and APU
  can proceed.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Acked-by: Tanmay Shah <tanmay.shah@amd.com>